Introduce a new Header.
