+++ src/components/Homepage/Header/styles.js
@@ -1,0 +1,23 @@
+import theme from '../../theme';
+
+const styles = {
+  link: {
+    ...theme.base,
+    color: '#333',
+    padding: '0 5px',
+    textDecoration: 'none',
+    fontSize: 15
+  },
+
+  container: {
+    textAlign: 'right',
+    padding: "60px 0",
+  },
+};
+
+styles.currentLink = {
+  ...styles.link,
+  fontWeight: 600,
+};
+
+export default styles;
